ETL_Project/requirements.txt
requests
pandas
psycopg2-binary
matplotlib